DjangoW/requirements.txt
Django==1.11.2
pip==9.0.1
pycryptodome==3.4.6
pytz==2017.2
setuptools==36.0.1
waitress==1.0.2
wheel==0.29.0
